feat: added SEO
