+++ src/app/(app)/purchases/[id]/page.tsx
@@ -2,6 +2,29 @@ import { AppLayout } from "@/components/layout/app-layout"
 import { getPurchaseByIdAction } from "@/app/_actions/purchase/getPurchaseByIdAction"
 import { PurchaseDetails } from "@/components/purchase/purchase-details"
 import { BackButton } from "@/components/misc/back-button"
+import type { Metadata } from "next"
+
+export async function generateMetadata({
+  params
+}: {
+  params: { id: string }
+}): Promise<Metadata> {
+  const purchase = await getPurchaseByIdAction(params.id)
+
+  if (!purchase) {
+    return {
+      title: "Compra não encontrada"
+    }
+  }
+
+  return {
+    title: `Detalhes da Compra #${purchase.id.slice(0, 8)} - eStore`,
+    description: `Veja os detalhes da sua compra #${purchase.id.slice(
+      0,
+      8
+    )} na eStore.`
+  }
+}
 
 export default async function PurchaseDetailsPage({
   params
